New helper objects (ParentInfo and ChildrenInfo) to capture details of parent / children in relationship - used in ParentChildrenSyncDownTarget to simplify code
